--- src/main/java/no/fint/linkwalker/Constants.java
@@ -1,7 +1,0 @@
-package no.fint.linkwalker;
-
-public enum Constants {
-    ;
-
-    public static final String PWF_BASE_URL = "https://play-with-fint.felleskomponent.no/";
-}
