--- src/main/java/no/fint/linkwalker/DiscoveredRelation.java
@@ -1,13 +1,0 @@
-package no.fint.linkwalker;
-
-import lombok.Data;
-
-import java.net.URL;
-import java.util.Set;
-
-@Data
-public class DiscoveredRelation {
-    private String rel;
-    private Set<URL> links;
-    private URL parentUrl;
-}
