--- src/main/java/no/fint/linkwalker/PwfUtils.java
@@ -1,12 +1,0 @@
-package no.fint.linkwalker;
-
-import org.apache.commons.lang3.StringUtils;
-
-public enum PwfUtils {
-    ;
-
-    public static boolean isPwf(String location) {
-        return StringUtils.startsWith(location, Constants.PWF_BASE_URL);
-    }
-
-}
