--- src/main/java/no/fint/linkwalker/TestedRelation.java
@@ -1,21 +1,0 @@
-package no.fint.linkwalker;
-
-import lombok.Data;
-import no.fint.linkwalker.dto.Status;
-
-import java.net.URL;
-
-@Data
-public class TestedRelation {
-
-    private final URL url;
-    private final URL parentUrl;
-    private Status status;
-    private String reason;
-
-    public TestedRelation(URL url, URL parentUrl) {
-        this.status = Status.RUNNING;
-        this.url = url;
-        this.parentUrl = parentUrl;
-    }
-}
