--- src/main/java/no/fint/linkwalker/dto/TestCaseViews.java
@@ -1,9 +1,0 @@
-package no.fint.linkwalker.dto;
-
-public class TestCaseViews {
-    public interface ResultsOverview {
-    }
-
-    public interface Details extends ResultsOverview {
-    }
-}
