--- src/main/java/no/fint/linkwalker/exceptions/FintLinkWalkerException.java
@@ -1,12 +1,0 @@
-package no.fint.linkwalker.exceptions;
-
-public class FintLinkWalkerException extends RuntimeException {
-
-    public FintLinkWalkerException(String message) {
-        super(message);
-    }
-
-    public FintLinkWalkerException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
