--- src/main/java/no/fint/linkwalker/exceptions/InvalidBaseUrlException.java
@@ -1,8 +1,0 @@
-package no.fint.linkwalker.exceptions;
-
-public class InvalidBaseUrlException extends FintLinkWalkerException {
-
-    public InvalidBaseUrlException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
