--- src/main/java/no/fint/linkwalker/exceptions/InvalidTestUrlException.java
@@ -1,12 +1,0 @@
-package no.fint.linkwalker.exceptions;
-
-public class InvalidTestUrlException extends FintLinkWalkerException {
-
-    public InvalidTestUrlException(String message) {
-        super(message);
-    }
-
-    public InvalidTestUrlException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
