+++ src/main/java/no/fintlabs/linkwalker/task/TaskController.java
@@ -1,0 +1,16 @@
+package no.fint.linkwalker.task;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.*;
+
+@Slf4j
+@RestController
+@RequestMapping("/tasks/{organization}")
+public class TaskController {
+
+    @PostMapping
+    public void postTask(@PathVariable String organization, @RequestBody Task task) {
+        log.info(task.toString());
+    }
+
+}
